third_party/WebKit/LayoutTests/fast/dom/shadow/nodetype-expected.txt

CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M54, around October 2016. See https://www.chromestatus.com/features/5716060992962560 for more details.

This tests the shadow host's and shadow root's nodeType.
PASS keygen.nodeType is Node.ELEMENT_NODE
PASS shadow.nodeType is Node.DOCUMENT_FRAGMENT_NODE
PASS shadowChild is non-null.
PASS shadowChild.nodeType != Node.DOCUMENT_FRAGMENT_NODE is true
PASS successfullyParsed is true

TEST COMPLETE
